+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.18.0] - 2022-09-05
+
 ### Added
 
 - Change `u64` and `VarInt` to `Amount` where it makes sense to by [@LeoNero](https://github.com/LeoNero) ([#113](https://github.com/monero-rs/monero-rs/pull/113))
@@ -259,7 +261,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Address and subaddress creation, de/serialization and validation
 - Private keys and one-time keys creation, de/serialization and validation
 
-[Unreleased]: https://github.com/monero-rs/monero-rs/compare/v0.17.2...HEAD
+[Unreleased]: https://github.com/monero-rs/monero-rs/compare/v0.18.0...HEAD
+[0.18.0]: https://github.com/monero-rs/monero-rs/compare/v0.17.2...v0.18.0
 [0.17.2]: https://github.com/monero-rs/monero-rs/compare/v0.17.1...v0.17.2
 [0.17.1]: https://github.com/monero-rs/monero-rs/compare/v0.17.0...v0.17.1
 [0.17.0]: https://github.com/monero-rs/monero-rs/compare/v0.16.0...v0.17.0